LPS-146974 Adds implementation, copy code from GetCollectionFieldMVCResourceCommand (no logic changes)
